Configure jest
Settings are from presentation slides of practice session 5.
"!src/store/actions/index.js" is not included because jest seems
to know that the file has no content to be tested (since it only
has export statements). We may need to test the file later if
that file grows to have some original lines of code.