feat: Add user profile summary page (#15)
This commit adds profile page summary tab to front-end.
The followings are some of the notable breaking changes:

* ProfileStatistics's tab buttons are moved to Profile
* actions.tsx is renamed to userActions.tsx
* userReducer's default export is renamed from reducer to userReducer
* Use User instead of UserField in types representing Redux state

To test the profile page, testing-library, which is shipped with
create-react-app, is used. We may need to add the library to "Testing
Plan" section in the design document.

To run the tests properly, jest is configured. Settings are from
presentation slides of practice session 5. "!src/store/actions/index.js"
is not included because jest seems to know that the file has no content
to be tested (since it only has export statements). We may need to test
the file later if that file grows to have some original lines of code.

This commit also adds not found page.

This commit adjusts package dependencies (closes PR #14):

* Downgrade eslint-plugin-jest

eslint-plugin-jest recently had a breaking change which is incompatible
with our linter config (eslint-config-react-app).

* Add @types/react-router-dom

To use components from react-router-dom, this package is needed.

* Downgrade history

We should use history version 4 for React Router version 5;
history version 5 is for React Router version 6.
See the following link: https://www.npmjs.com/package/history

Resolves: #9
See Also: #3