Change Travis CI settings
This commit trys to fix the failing build. The build is failing because
ESLint complains that 'babel-eslint' is required for
'eslint-config-react-app'. I found out that while VS Code runs ESLint
just fine, running ESLint from command line results in a crash.

Since babel-eslint is no longer maintained, I did not include it in
dependency originally. However, it seems that babel-eslint is essential
for us to use ESLint in Travis CI.

This commit also changes yarn command to include flag --frozen-lockfile
because it is included in default install command (see link
https://docs.travis-ci.com/user/languages/javascript-with-nodejs/).

Finally, this commit changes Python and node versions. I personally
think we should use the closest version we use in development, when we
configure the language versions in CI.