Update .travis.yml
Travis failed again because it could not locate SonarCloud configuration file (see the build log). The reason seems to be that sonar-scanner is not being executed on the root directory. Therefore, this commit fixes that. Also, this commit corrects organization key to swsnu (I don't know why I did not notice it earlier).

In hindsight, maybe we should not have put "probloom" directory inside frontend and backend. Sorry for that.